Refactor : move followerLogic to FollowRepository - #49
