+++ feature/home/src/main/java/com/kolown/home/HomeViewModel.kt
@@ -4,6 +4,7 @@ import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import androidx.lifecycle.viewmodel.compose.viewModel
+import com.kolown.data.repository.FollowRepository
 import com.kolown.data.repository.PostRepository
 import com.kolown.model.PostContentModel
 import com.kolown.model.Reactions
@@ -23,6 +24,7 @@ import javax.inject.Inject
 @HiltViewModel
 class HomeViewModel @Inject constructor(
     private val postRepository: PostRepository,
+    private val followRepository: FollowRepository
 ) : ViewModel() {
     private val _uiState = MutableStateFlow<UiState<List<PostContentModel>>>(UiState.Loading)
     val uiState = _uiState.asStateFlow()
@@ -35,7 +37,7 @@ class HomeViewModel @Inject constructor(
 
     fun followUser(id: String, name: String) {
         updateFollow(id)
-        postRepository.followUser(id, name)
+        followRepository.followUser(id, name)
             .catch { Log.e("FollowUpload", "viewModel: $it") }
             .launchIn(viewModelScope)
     }
@@ -43,7 +45,7 @@ class HomeViewModel @Inject constructor(
     fun unFollowUser(id: String) {
         viewModelScope.launch {
             updateFollow(id)
-            postRepository.unFollowUser(id)
+            followRepository.unFollowUser(id)
                 .catch { Log.e("UnFollowUpload", "viewModel: $it") }
                 .launchIn(viewModelScope)
         }
